added accept and reject applications

+++ src/main/java/com/example/apr/security/AuthUser.java
@@ -14,6 +14,6 @@ public class AuthUser {
     private String firstName;
     private String lastName;
     private String role;
-    private long companyId;
+    private int companyId;
 
 }
